+++ src/main/webapp/WEB-INF/views/notice.jsp
@@ -43,7 +43,7 @@ $(document).ready(function(){
 <div class="content-body">
 	<div class="container-fluid">
 		<!-- 공지사항 전체 리스트 출력 -->
-		<form id="search_form" action="/edu/notice" method="post" class="form-inline d-flex justify-content-end">
+		<form id="search_form" action="/notice" method="post" class="form-inline d-flex justify-content-end">
 			<input type="hidden" name="empNo" value="${sessionEmp.empNo }">
 			<input type="hidden" name="a" value="list"> 
 			<select name="keyField" size="1" id="keyField" class="form-control form-control-sm">
@@ -86,7 +86,7 @@ $(document).ready(function(){
 			<nav aria-label="Page navigation example">
 			<ul class="pagination justify-content-center">
 				<c:if test="${paging.hasPrev }">
-					<li class="page-item"><a class="page-link" href="/edu/notice?pageNum=${paging.startPage-1}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">이전</a></li>
+					<li class="page-item"><a class="page-link" href="/notice?pageNum=${paging.startPage-1}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">이전</a></li>
 				</c:if>
 				
 				<c:forEach var="p" begin="${paging.startPage }" end="${paging.endPage }" step="1">
@@ -97,10 +97,10 @@ $(document).ready(function(){
 						<c:otherwise>
 							<c:choose>
 								<c:when test="${empty pageInfo.keyField}">
-									<li class="page-item"><a class="page-link" href="/edu/notice?pageNum=${p}">${p}</a>
+									<li class="page-item"><a class="page-link" href="/notice?pageNum=${p}">${p}</a>
 								</c:when>
 								<c:otherwise>
-									<li class="page-item"><a class="page-link" href="/edu/notice?pageNum=${p}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">${p}</a>
+									<li class="page-item"><a class="page-link" href="/notice?pageNum=${p}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">${p}</a>
 								</c:otherwise>
 							</c:choose>
 							
@@ -109,7 +109,7 @@ $(document).ready(function(){
 				</c:forEach>
 
 				<c:if test="${paging.hasNext }">
-					<li class="page-item"><a class="page-link" href="/edu/notice?pageNum=${paging.endPage+1}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">다음</a></li>
+					<li class="page-item"><a class="page-link" href="/notice?pageNum=${paging.endPage+1}&keyField=${pageInfo.keyField }&keyword=${pageInfo.keyword}">다음</a></li>
 				</c:if>
 			</ul>
 			</nav>
